Move testcase properties parser to xunit2
